Split fluent builder generator into its own class
